cofilab-backend/requirements.txt
Django>=4.2,<5
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
channels
channels-redis
django-cors-headers
celery
redis
python-decouple
cryptography
requests
